TC01-Docker-RestAPI/pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = ["Steven Sequeira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"
passlib = "^1.7.4"
jwt = "^1.3.1"
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
